Can now register descriptors in page maps
Additionally, change type of superblock and heap to raw pointers.
Having superblock is understandable - it is literally a reference to
memory, so we should have it as a ptr.
For heap, it is NULL if descriptor points to an oversized superblock. We
cannot have it as an Option - then the Descriptor would own the heap.
&Option also does not make sence as we would have to change HEAPS
global.