Change locking to be coarse-grained
This means that we now have a singe lock (POOL_LOCK) that is acquired
when the list of Descriptors is being modified. There are four functions
that do this: Descriptor::{alloc, retire}, heap_pop_partial and
heap_push_partial.

Additionally, change inner type of partial list of descriptors inside of
ProcHeap to be RefCell. This prevents mutability conflicts.