Optimize apf analysis further; make hooks for user-controlled APF value
This changes (most) values used for apf analysis to u32 instead of
usize, which reduces the memory footprint of the analysis.

Hooks for user-controlled APF values take size (presumably of the
allocated data structure), but set APF for the corresponding size class,
thus APF values are set per-thread per-size class.